split chart into two separate components
