+++ src/components/Chart.jsx
@@ -4,6 +4,7 @@ import HC_exporting from "highcharts/modules/exporting";
 import highchartsAccessibility from "highcharts/modules/accessibility";
 import { useState, useRef } from "react";
 import Papa from "papaparse";
+import ChartRenderer from "./ChartRenderer";
 
 // add the exporting and accessibility modules
 if (typeof HighCharts === "object") {
@@ -25,6 +26,13 @@ function Chart() {
   const [padjThreshold, setPadjThreshold] = useState(0);
   const [lowerLog2FCThreshold, setLowerLog2FCThreshold] = useState(0);
   const [higherLog2FCThreshold, setHigherLog2FCThreshold] = useState(0);
+
+  // State to store plotLine values
+  const [plotLines, setPlotLines] = useState({
+    padjThresholdLine: null,
+    lowerLog2FCThresholdLine: null,
+    higherLog2FCThresholdLine: null,
+  });
 
   const handleDataParsing = (data) => {
     const significantDataTemp = [];
@@ -62,6 +70,31 @@ function Chart() {
     setNoChangeCount(noChangeCountTemp);
   };
 
+  const updatePlotLines = () => {
+    // Update plotLines based on threshold values
+    const newPlotLines = {
+      padjThresholdLine: {
+        value: padjThreshold,
+        color: "grey",
+        dashStyle: "shortdash",
+        width: 1,
+      },
+      lowerLog2FCThresholdLine: {
+        value: lowerLog2FCThreshold,
+        color: "grey",
+        dashStyle: "shortdash",
+        width: 1,
+      },
+      higherLog2FCThresholdLine: {
+        value: higherLog2FCThreshold,
+        color: "grey",
+        dashStyle: "shortdash",
+        width: 1,
+      },
+    };
+    setPlotLines(newPlotLines);
+  };
+
   // use the papa parse package to parse the csv file
   const handleSubmit = (event) => {
     event.preventDefault();
@@ -71,100 +104,10 @@ function Chart() {
       dynamicTyping: true,
       complete: function (results) {
         handleDataParsing(results.data);
+        updatePlotLines();
         setShowChart(true);
       },
     });
-  };
-
-  // set the options for the chart
-  const options = {
-    chart: {
-      type: "scatter",
-      zoomType: "xy",
-    },
-    title: {
-      text: "Volcano Plot",
-    },
-    subtitle: {
-      text: "Click and drag within the plot area to zoom in",
-    },
-    xAxis: {
-      zIndex: 10,
-      title: {
-        text: "log2FoldChange",
-      },
-      plotLines: [
-        {
-          value: lowerLog2FCThreshold,
-          color: "grey",
-          dashStyle: "shortdash",
-          width: 1,
-        },
-        {
-          value: higherLog2FCThreshold,
-          color: "grey",
-          dashStyle: "shortdash",
-          width: 1,
-        },
-      ],
-    },
-    yAxis: {
-      tickInterval: 25,
-      lineWidth: 1,
-      title: {
-        text: "-log10(padj)",
-      },
-      min: -10,
-      plotLines: [
-        {
-          value: padjThreshold,
-          color: "grey",
-          dashStyle: "shortdash",
-          width: 1,
-        },
-      ],
-      startOnTick: false,
-    },
-    plotOptions: {
-      series: {
-        // increase the turboThreshold to
-        // allow for more data points
-        turboThreshold: 500000,
-        marker: {
-          symbol: "circle",
-          radius: 2,
-        },
-      },
-    },
-    series: [
-      {
-        name: "Significant Genes",
-        color: "rgb(133,31,76)",
-        data: significantData,
-      },
-      {
-        name: "No Change",
-        color: "rgb(150,156,159)",
-        data: notSignificantData,
-      },
-    ],
-    tooltip: {
-      formatter: function () {
-        return `${this.point.gene} <br> <b>log2FC:</b> ${this.point.x.toFixed(
-          2
-        )} <br> <b>-log10(padj):</b> ${this.point.y.toFixed(2)}`;
-      },
-    },
-    // add a custom legend to show the number of
-    // data points for each series
-    legend: {
-      labelFormatter: function () {
-        const series = this;
-        const numPoints =
-          series.name === "Significant Genes" ? changeCount : noChangeCount;
-        return `${series.name} (${numPoints})`;
-      },
-    },
   };
 
   return (
@@ -193,54 +136,61 @@ function Chart() {
             gap: "20px",
           }}
         >
-          <label>-Log10(padj) Threshold:</label>
-          <input
-            type="range"
-            min={0}
-            max={2}
-            step={0.01}
-            value={padjThreshold}
-            onChange={(e) => setPadjThreshold(e.target.value)}
-            data-testid="padj-threshold"
-          />
-          <span>{padjThreshold}</span>
-          <br />
-          <label>Lower Log2FC Threshold:</label>
-          <input
-            type="range"
-            min={-4}
-            max={0}
-            step={0.1}
-            value={lowerLog2FCThreshold}
-            onChange={(e) => setLowerLog2FCThreshold(e.target.value)}
-            data-testid="lower-log2FC-threshold"
-          />
-          <span>{lowerLog2FCThreshold}</span>
-          <br />
-          <label>Higher Log2FC Threshold:</label>
-          <input
-            type="range"
-            min={0}
-            max={4}
-            step={0.1}
-            value={higherLog2FCThreshold}
-            onChange={(e) => setHigherLog2FCThreshold(e.target.value)}
-            data-testid="higher-log2FC-threshold"
-          />
-          <span>{higherLog2FCThreshold}</span>
+          <div style={{ width: 400 }}>
+            <label>-Log10(padj) Threshold:</label>
+            <br />
+            <input
+              type="range"
+              min={0}
+              max={2}
+              step={0.01}
+              value={padjThreshold}
+              onChange={(e) => setPadjThreshold(e.target.value)}
+              data-testid="padj-threshold"
+            />
+            <span>{padjThreshold}</span>
+          </div>
+          <div style={{ width: 400 }}>
+            <label>Lower Log2FC Threshold:</label>
+            <br />
+            <input
+              type="range"
+              min={-4}
+              max={0}
+              step={0.1}
+              value={lowerLog2FCThreshold}
+              onChange={(e) => setLowerLog2FCThreshold(e.target.value)}
+              data-testid="lower-log2FC-threshold"
+            />
+            <span>{lowerLog2FCThreshold}</span>
+          </div>
+          <div style={{ width: 400 }}>
+            <label>Higher Log2FC Threshold:</label>
+            <br />
+            <input
+              type="range"
+              min={0}
+              max={4}
+              step={0.1}
+              value={higherLog2FCThreshold}
+              onChange={(e) => setHigherLog2FCThreshold(e.target.value)}
+              data-testid="higher-log2FC-threshold"
+            />
+            <span>{higherLog2FCThreshold}</span>
+          </div>
         </div>
         <button type="submit" data-testid="submit-button">
           Submit
         </button>
       </form>
       {showChart && (
-        <div
-          className="chart"
-          style={{ width: 500, height: 300, margin: "50px auto" }}
-          data-testid="chart"
-        >
-          <HighChartsReact highcharts={HighCharts} options={options} />
-        </div>
+        <ChartRenderer
+          significantData={significantData}
+          notSignificantData={notSignificantData}
+          plotLines={plotLines}
+          changeCount={changeCount}
+          noChangeCount={noChangeCount}
+        />
       )}
     </div>
   );
